# fix a bunch of outstanding issues