--- a/.clang-format
+++ b/.clang-format
@@ -7,6 +7,7 @@ AlignEscapedNewlines: DontAlign
 AlwaysBreakBeforeMultilineStrings: true
 AlwaysBreakTemplateDeclarations: false
 AllowAllParametersOfDeclarationOnNextLine: false
+AllowShortFunctionsOnASingleLine: false
 BreakBeforeBraces: Linux
 IndentWidth:     4
 KeepEmptyLinesAtTheStartOfBlocks: false