--- a/go.mod
+++ b/go.mod
@@ -3,7 +3,7 @@ module github.com/cloudflare/rakelimit
 go 1.14
 
 require (
-	github.com/cilium/ebpf v0.0.0-20200917151652-9b4cff7de01e
+	github.com/cilium/ebpf v0.3.0
 	github.com/google/gopacket v1.1.18
 	golang.org/x/sys v0.0.0-20200124204421-9fbb57f87de9
 )