--- a/src/countmin.h
+++ b/src/countmin.h
@@ -32,15 +32,15 @@ struct countmin {
 };
 
 // add element and determine count
-static __u64 FORCE_INLINE add_to_cm(struct countmin *cm, __u64 ts, struct packet_element *element)
+static __u64 FORCE_INLINE add_to_cm(struct countmin *cm, __u64 now, struct packet_element *element)
 {
 	fpoint min = -1;
 #pragma clang loop unroll(full)
 	for (int i = 0; i < HASHFN_N; i++) {
 		__u32 target_idx       = fasthash64(element, sizeof(struct packet_element), i) & (COLUMNS - 1);
 		struct cm_value *value = &cm->values[i][target_idx];
-		value->value           = estimate_avg_rate(value->value, ts - value->ts);
-		value->ts              = ts;
+		value->value           = estimate_avg_rate(value->value, value->ts, now);
+		value->ts              = now;
 		if (value->value < min) {
 			min = value->value;
 		}