--- a/src/ewma.h
+++ b/src/ewma.h
@@ -5,15 +5,17 @@
 #include "common.h"
 #include "fixed-point.h"
 
-// 100ms
-#define WINDOW_NS 100000000ull
-#define ONE_SECOND_NS 1000000000ull
-
 // estimate_avg_rate takes a previous rate and a duration that elapsed
 // since this rate has been determined, and estimates based on these and
 // WINDOW the current rate in packets per second.
 static __u32 FORCE_INLINE estimate_rate(__u32 old_rate, __u64 old_ts, __u64 now)
 {
+	// The window after which old observations are discarded.
+	// Chosen to be a power of two so that division can be done
+	// with a bit shift.
+	const __u32 WINDOW_NS     = 1ull << 27;
+	const __u32 ONE_SECOND_NS = 1000000000ull;
+
 	if (old_ts >= now) {
 		// Time went backward or stood still due to clockskew. Return the old value,
 		// since we can't compute the current rate.
@@ -27,7 +29,7 @@ static __u32 FORCE_INLINE estimate_rate(__u32 old_rate, __u64 old_ts, __u64 now)
 		return 0;
 	}
 
-	__u32 rate_current = ONE_SECOND_NS / elapsed;
+	__u32 rate_current = ONE_SECOND_NS / (__u32)elapsed;
 	if (old_rate == 0) {
 		// This is the first time we can calculate a rate, so use that
 		// to initialize our estimate.