--- a/src/ewma.h
+++ b/src/ewma.h
@@ -12,21 +12,29 @@
 // estimate_avg_rate takes a previous rate and a duration that elapsed
 // since this rate has been determined, and estimates based on these and
 // WINDOW the current rate in packets per second.
-static fpoint FORCE_INLINE estimate_avg_rate(fpoint old_rate, __s64 dur)
+static fpoint FORCE_INLINE estimate_avg_rate(fpoint old_rate, __u64 old_ts, __u64 now)
 {
-	// if nothing changed or
-	if (dur <= 0) {
+	__s64 elapsed = now - old_ts;
+	if (elapsed <= 0) {
+		// Time went backward or stood still due to clockskew. Return the old value,
+		// since we can't compute the current rate.
 		return old_rate;
 	}
 
-	// calculate pps since last timestamp
-	fpoint rate_current = to_fixed_point(ONE_SECOND_NS / dur, 0);
-	// if the last timestamp is older than the window the new pps will simply be rate_current
-	if (dur >= WINDOW_NS) {
+	if (old_ts == 0 || elapsed >= WINDOW_NS) {
+		// Either there is no previous measurement, or it's too old.
+		// We need another sample to calculate a reliable rate.
+		return 0;
+	}
+
+	fpoint rate_current = to_fixed_point(ONE_SECOND_NS / elapsed, 0);
+	if (old_rate == 0) {
+		// This is the first time we can calculate a rate, so use that
+		// to initialize our estimate.
 		return rate_current;
 	}
 
-	fpoint a = to_fixed_point(dur, 0) / WINDOW_NS;
+	fpoint a = to_fixed_point(elapsed, 0) / WINDOW_NS;
 
 	fpoint new_rate = old_rate;
 	if (old_rate > rate_current) {