--- a/src/fixed-point.h
+++ b/src/fixed-point.h
@@ -10,9 +10,17 @@
 
 typedef __u64 fpoint;
 
-static __u64 FORCE_INLINE to_fixed_point(__u32 n) { return ((__u64)n) << FRACTION_BITS; }
+static __u64 FORCE_INLINE to_fixed_point(__u32 integer, __u32 fraction)
+{
+	return (((__u64)integer) << FRACTION_BITS) | (__u64)fraction;
+}
 
-static __u32 FORCE_INLINE to_int(fpoint a) { return a >> FRACTION_BITS; }
+static __u32 FORCE_INLINE to_int(fpoint a)
+{
+	return a >> FRACTION_BITS;
+}
 
-typedef __u64 fpoint;
-static fpoint FORCE_INLINE div_by_int(fpoint a, fpoint b) { return a / to_int(b); }
+static fpoint FORCE_INLINE div_by_int(fpoint dividend, __u32 divisor)
+{
+	return dividend / divisor;
+}