rename test file
    
    Rename scenarios_test.go now that there are no scenarios anymore.

remove scenarios from tests
    
    We now have more specific tests that cover what the scenarios test.
    Remove the scenarios.

bpf: fix IPv6 and add tests
    
    Fix IPv6 support and add tests to make sure we don't regress. The semantics of
    IPv6 addresses had to be changed to look at /128, otherwise the tests don't
    work for both IPv4 and IPv6. It's not clear that /64 makes more sense than
    /128 so I decided to keep it simple.
    
    Fixes #8

bpf: replace division with bit shift
    
    Divisions are quite expensive, so replace the division by the window with
    a bit shift. We can achieve this by choosing a window that is a power of
    two.
    
        name              old time/op  new time/op  delta
        Rakelimit/IPv4-8   672ns ± 2%   528ns ± 2%  -21.41%  (p=0.029 n=4+4)

return a helpful error when optmem sysctl is too low
    
    Make it easier for the user to figure out what to do when the sysctl is
    too low for the socket filter.

bpf: use two distinct hash functions
    
    Varying the seed to fasthash doesn't buy us additional collision resistance.
    Consider two distinct buffers with the same hash:
    
        fasthash(A, 0) == fasthash(B, 0)
    
    Changing the seed doesn't resolve the collision:
    
        fasthash(A, 42) == fasthash(B, 42)
    
    On the other hand, using a distinct hash function it's likely that
    
        hashlittle(A, 0) != hashlittle(B, 0)
    
    Due to the properties of the count-min sketch this leads to a better
    estimate.

bpf: check that packets are dropped on the correct level
    
    For testing, store the level a packet was dropped on in the argument
    map and check this in tests.

bpf: fix ewma calculation
    
    The current EWMA calculation has some sort of weird overflow. Use the
    canonical form which doesn't suffer from this problem.

bpf: reduce countmin value size to 32bit
    
    We don't really need 64 bit to represent a rate, 2^32 packets per
    second should be enough for everyone. Once we manage to shrink the
    timestamp to 32 bit we'll save a bunch of memory.
    
    Updates #16

bpf: specialise for IP version
    
    There is an important optimization to be had: since we know the IP
    version of a socket we can attach a socket filter that is specialised
    to IPv4 or IPv6. This is a big win for IPv4, since we end up hashing
    a lot less data.
    
    We also generate less code, so we are closer to not needing a higher
    optmem limit for IPv4.
    
    Fixes #2
    Updates #13

bpf: encode generalisations in a table
    
    Move generalisations into a table. This makes it much easier to add
    a new one and is also easier to read. We rely on clang to inline the
    loop.

bpf: turn element into void pointer and length
    
    Make the countmin sketch take a pointer and a length instead of a
    concrete type. This allows us to specialise the socket filter to
    different IP versions later.

Update github.com/cilium/ebpf

bpf: test attack rate propagation behaviour
    
    Add a test that ensures we don't propagate attack traffic to higher
    levels, since this would lead to dropping non-attack traffic.

bpf: test each generalisation separately
    
    Test each generalisation that we know about, to make sure they are
    working correctly. Turns out that some level 3 generalisations don't
    work, so fix them.

bpf: add a test for accurate rate limit
    
    So far, accurate rate limiting is only tested as a side effect of the
    scenarios we run. Add a dedicated test, which uses the dumbest "attack"
    to reduce the amount of moving parts.

refactor tests
    
    Allow tests to override both current time and the random value. Also
    make some of the fixed point functions easier to understand. Use a
    different seed on each run of the test suite to increase coverage.