@lmb 2021-01-19
@sauercrowd let me know if you're at all up for reviewing this, it's quite a code dump ;P